feature : 회원가입 기능구현
mongoDB의 API를 도와주는 helper 함수 구현
회원가입 기능 구현
- 이미 가입된 회원인 경우 오류를 보냄
