+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "workbench.colorCustomizations": {
+    "titleBar.activeBackground": "#ff0000"
+  }
+}
