+++ package.json
@@ -17,6 +17,7 @@
     "antd": "^5.6.4",
     "axios": "^1.4.0",
     "eslint-config-next": "13.4.7",
+    "mongodb": "^5.7.0",
     "next": "13.4.7",
     "react": "18.2.0",
     "react-dom": "18.2.0",
