--- pages/api/create.js
@@ -1,3 +1,0 @@
-export default function handler(req, res) {
-  res.status(200).json({ name: "처리완료" });
-}
